% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapStatistics.R
\name{gapStatistics}
\alias{gapStatistics}
\title{gap statistics}
\usage{
gapStatistics(d, K = 3, B = 10, gamma = NULL, alpha = 1, group = NULL,
  seed = 15213, silence = FALSE)
}
\arguments{
\item{d}{A list of S studies, each study is a combined data matrix n*J, where n is number of subjects, J=J1+J2+... and J1 is number of features in omics dataset 1 and J2 is number of features in omics dataset 2...}

\item{K}{number of clusters}

\item{B}{number of permutations.}

\item{gamma}{Penalty on total number of features. Default is given. Larger gamma will yeild small number of selected features.}

\item{alpha}{balance between group sparsity and individual sparsity. alpha=1 yeilds no group sparsity. alpha=0 yeilds no individual penalty.}

\item{group}{Prior group information. Potentially these group can contain overlap features. group is a list and each element of the list is feature index.}

\item{seed}{random seed}

\item{silence}{do not print progress in silence = TRUE.}
}
\value{
a table. Each row represents a gamma. 
\item{gamma}{input gammas}
\item{score}{objective score: see obj0 in ISKmeans function}
\item{E.score}{mean value of permutated score}
\item{se.score}{standard error of permutated score}
\item{gapStat}{gap statistics, score - E.score}
\item{numF}{number of selected features}
}
\description{
gap statistics to tune gamma
}
\details{
gap statistics to tune gamma, the tuning parameter to control number of features.
}
\examples{

S <- 2
K <- 3
G <- 1000
g1 <- 50
g2 <- 50
n0 <- 20
n <- K*n0
labels <- cut(1:n,breaks=K,labels=FALSE)

set.seed(32611)
S1 <- matrix(rnorm(G*n), nrow=G, ncol=n)
S2 <- matrix(rnorm(G*n), nrow=G, ncol=n)

S1[1:g1, labels==1] <- S1[1:g1, labels==1] + 2
S1[1:g1, labels==3] <- S1[1:g1, labels==3] - 2
S1[g1 + 1:g2, labels==1] <- S1[g1 + 1:g2, labels==1] - 2
S1[g1 + 1:g2, labels==2] <- S1[g1 + 1:g2, labels==2] + 2

S2[1:g1, labels==2] <- S2[1:g1, labels==2] + 2
S2[1:g1, labels==1] <- S2[1:g1, labels==1] - 2
S2[g1 + 1:g2, labels==2] <- S2[g1 + 1:g2, labels==2] - 2
S2[g1 + 1:g2, labels==3] <- S2[g1 + 1:g2, labels==3] + 2

S = list(t(S1),t(S2))
groups <- Map('c',1:g1,g1 + 1:g2)

gapResult <- gapStatistics(d=S,K=3,B=10,group=group)
print(gapResult)
}
\author{
Caleb
}
