% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferLambda.R
\name{inferLambda}
\alias{inferLambda}
\title{Infer lambda}
\usage{
inferLambda(resMISKmeans, balance = 1/2)
}
\arguments{
\item{resMISKmeans}{The result returned from MISKmeans.}

\item{balance}{selecting lambda such that the separation ability = balance * matching ability}
}
\value{
lambda values
}
\description{
MISKmenas
}
\details{
Estimate a propriate tuning parameter lambda
}
\examples{

S <- 2
K <- 3
G <- 1000
g1 <- 50
g2 <- 50
n0 <- 20
n <- K*n0
labels <- cut(1:n,breaks=K,labels=FALSE)

set.seed(32611)
S1 <- matrix(rnorm(G*n), nrow=G, ncol=n)
S2 <- matrix(rnorm(G*n), nrow=G, ncol=n)

S1[1:g1, labels==1] <- S1[1:g1, labels==1] + 2
S1[1:g1, labels==3] <- S1[1:g1, labels==3] - 2
S1[g1 + 1:g2, labels==1] <- S1[g1 + 1:g2, labels==1] - 2
S1[g1 + 1:g2, labels==2] <- S1[g1 + 1:g2, labels==2] + 2

S2[1:g1, labels==2] <- S2[1:g1, labels==2] + 2
S2[1:g1, labels==1] <- S2[1:g1, labels==1] - 2
S2[g1 + 1:g2, labels==2] <- S2[g1 + 1:g2, labels==2] - 2
S2[g1 + 1:g2, labels==3] <- S2[g1 + 1:g2, labels==3] + 2

S = list(t(S1),t(S2))
groups <- Map('c',1:g1,g1 + 1:g2)

res <- MISKmeans(d = S, K = 3, gamma = 0.4, group = groups)

inferLambda(res) 

}
\author{
Caleb
}
