% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MISKmeans.R
\name{MISKmeans}
\alias{MISKmeans}
\title{MISKmeans}
\usage{
MISKmeans(d, K = NULL, gamma = NULL, lambda = 0.5, alpha = 0.5,
  group = NULL, nstart = 20, wsPre = NULL, iniWbound = 20,
  penaltyInfo = NULL, silent = FALSE, maxiter = 20,
  sampleSizeAdjust = FALSE)
}
\arguments{
\item{d}{combined data matrix n*J, where n is number of subjects, J=J1+J2+... and J1 is number of features in omics dataset 1 and J2 is number of features in omics dataset 2...}

\item{K}{number of clusters}

\item{gamma}{Penalty on total number of features. Larger gamma will yeild small number of selected features.}

\item{lambda}{A tuning parameter controlling the balance between separation ability (BCSS/TSS) and matching function. lambda is set to be 0.5 by default.}

\item{alpha}{balance between group sparsity and individual sparsity. alpha=1 yeilds no group sparsity. alpha=0 yeilds no individual penalty. Default alpha=0.5}

\item{group}{Prior group information. Potentially these group can contain overlap features. group is a list and each element of the list is feature index.}

\item{nstart}{Number of initials for Kmeans for sparse Kmeans}

\item{wsPre}{Initial feature weight.}

\item{penaltyInfo}{only for the purpose of gap statitics. Here we will fix the penalty design to perform gap statistics. The input should be a list of groupInfo. See groupInfo for details.}

\item{silent}{Output progress.}

\item{maxiter}{Maximum numbre of iteration between ws and Cs.}

\item{sampleSizeAdjust}{logical argument,
controlling whether to adjust for sample size.
If true, that means study with larger sample size will have a larger impact.
If false, each study has equal contribution.
Without prior information, sampleSizeAdjust=FALSE is suggested since we are not sure about data quality.}
}
\value{
m lists, m is length of gamma parameters. If gamma is a scalar, m = 1. The following items are included in the list.
\item{ws}{weight for each feature. Zero weight means the feature is not selected.}
\item{Cs}{Cluster Assignment}
\item{obj0}{sum of weighted separation ability. This term is for the purpose of gap statistics.}
\item{objective}{objective value}
\item{groupInfo}{a list containing group design, alpha, gamma}
}
\description{
MISKmenas
}
\details{
Meta-analytic multi-level omics data integration with incorporation of prior group structure
}
\examples{

S <- 2
K <- 3
G <- 1000
g1 <- 50
g2 <- 50
n0 <- 20
n <- K*n0
labels <- cut(1:n,breaks=K,labels=FALSE)

set.seed(32611)
S1 <- matrix(rnorm(G*n), nrow=G, ncol=n)
S2 <- matrix(rnorm(G*n), nrow=G, ncol=n)

S1[1:g1, labels==1] <- S1[1:g1, labels==1] + 2
S1[1:g1, labels==3] <- S1[1:g1, labels==3] - 2
S1[g1 + 1:g2, labels==1] <- S1[g1 + 1:g2, labels==1] - 2
S1[g1 + 1:g2, labels==2] <- S1[g1 + 1:g2, labels==2] + 2

heatmap(S1, Rowv=NA, Colv=NA)

S2[1:g1, labels==2] <- S2[1:g1, labels==2] + 2
S2[1:g1, labels==1] <- S2[1:g1, labels==1] - 2
S2[g1 + 1:g2, labels==2] <- S2[g1 + 1:g2, labels==2] - 2
S2[g1 + 1:g2, labels==3] <- S2[g1 + 1:g2, labels==3] + 2

heatmap(S2, Rowv=NA, Colv=NA)

S = list(t(S1),t(S2))
groups <- Map('c',1:g1,g1 + 1:g2)

res <- MISKmeans(d = S, K = 3, gamma = 0.4, group = groups)

names(res)

order_S1 <- order(res$Cs[[1]])
order_S2 <- order(res$Cs[[2]])

S1_sub <- S1[res$ws!=0, order_S1]
S2_sub <- S2[res$ws!=0, order_S2]

col_S1 <- palette()[res$Cs[[1]][order_S1]]
col_S2 <- palette()[res$Cs[[2]][order_S2]]

heatmap(S1_sub, Rowv=NA, Colv=NA, ColSideColors=col_S1)
heatmap(S2_sub, Rowv=NA, Colv=NA, ColSideColors=col_S2)



}
\author{
Caleb
}
